opted to reuse the algorithm for multiple linear regression for multivariate linear regression

+++ Main.py
@@ -119,8 +119,9 @@ def testMultivariateReg():
     
     betaMatrix = MultivariateLinearRegression(xMatrix,yMatrix)
 
-    print(betaMatrix)
-    #def reg
+    print(betaMatrix.astype(float))
+    
+    
 
 #--------------------------------------------------------------------------------------------------------------------------------------------------------
